separated component from registration
